+++ include/boost/fusion/support/detail/as_fusion_element.hpp
@@ -9,7 +9,7 @@
 #define FUSION_AS_FUSION_ELEMENT_05052005_0338
 
 #include <boost/fusion/support/config.hpp>
-#include <boost/ref.hpp>
+#include <boost/core/ref.hpp>
 
 #ifndef BOOST_NO_CXX11_HDR_FUNCTIONAL
 #include <functional>
